New bigger version field and new api routes

+++ src/infrastructure/schemas/MosaicAliasTransactionSchema.ts
@@ -34,11 +34,11 @@ export default new Schema([
     uint('size'),
     array('signature'),
     array('signer'),
-    ushort('version'),
+    uint('version'),
     ushort('type'),
     array('fee', TypeSize.INT),
     array('deadline', TypeSize.INT),
     ubyte('actionType'),
     array('namespaceId', TypeSize.INT),
-    array('mosaicId', TypeSize.INT)
+    array('mosaicId', TypeSize.BYTE)
 ]);
