+++ src/infrastructure/schemas/SecretLockTransactionSchema.ts
@@ -34,7 +34,7 @@ const schema = new Schema([
     uint('size'),
     array('signature'),
     array('signer'),
-    ushort('version'),
+    uint('version'),
     ushort('type'),
     array('fee', TypeSize.INT),
     array('deadline', TypeSize.INT),
